+++ lib/beamer/locations/settings_location.dart
@@ -22,6 +22,7 @@ import 'package:lotti/pages/settings/measurables/measurable_details_page.dart';
 import 'package:lotti/pages/settings/measurables/measurables_page.dart';
 import 'package:lotti/pages/settings/outbox/outbox_monitor.dart';
 import 'package:lotti/pages/settings/settings_page.dart';
+import 'package:lotti/pages/settings/speech/speech_settings_page.dart';
 import 'package:lotti/pages/settings/sync/sync_assistant_page.dart';
 import 'package:lotti/pages/settings/tags/create_tag_page.dart';
 import 'package:lotti/pages/settings/tags/tag_edit_page.dart';
@@ -211,6 +212,13 @@ class SettingsLocation extends BeamLocation<BeamState> {
         const BeamPage(
           key: ValueKey('settings-health_import'),
           child: HealthImportPage(),
+        ),
+
+      // Speech Settings
+      if (pathContains('speech_settings'))
+        const BeamPage(
+          key: ValueKey('speech_settings'),
+          child: SpeechSettingsPage(),
         ),
 
       // Advanced Settings